+++ src/components/location/InterestSelector.tsx
